Day-8/requirements.txt
fastapi
uvicorn
authlib
httpx
python-dotenv
itsdangerous
jinja2
python-jose[cryptography]
passlib[bcrypt]
pytest
pytest-asyncio
pytest-cov
